examples/src/main/python/requirements.txt
grpcio==1.34.1
google-api-python-client==1.12.8
grpcio-tools==1.34.1